tidy up of attachment integration test; asset loader
#29
semver:patch
